-n Fixed nova-novncproxy.
Before this commit, nova-novncproxy would not connect. It turns
out that if memcached is configured all nova compute nodes
must be able to see the cache. After this commit memcached was
placed on the management network with the appropriate firewall
rules, consoleauth connections were being completed, and the
browser could create vnc connections to the system guests.