1.1.0 release
This releases fixes many bugs related to the ordering of resources
in this module. In particular, the firewall rules are applied
after repositories and packages have been installed, preventing
the accidental locking out of a machine from required network
access. Additionally, the module was bumped to depend on the
latest puppetlabs-vswitch release, allowing for more reliable
creation of the external ovs bridge.