Update tempest stuff for multi-node
The puppetdb module needs a 3.x version to work.

One issue is that the tempest_glance_id_setter resource uses the
glance_image resource type, which expects an /etc/glance/glance.conf
file to exist. Since this doesn't exist on the controller, it doesn't
work yet.

Another issue is that the public_network_id needs to be added to the
tempest.conf, but can't until the public network exists. So it has to be
run again after the network node is configured.

This at least gets the code into the tree.