Updated Glance to put glance-api.conf on controller
Tempest, when run from the controller, needs the *.conf
files to set up authentication tokens. Previously glance-api.conf
was not on the controller node, causing Tempest errors in relation
to working with images. This patch generalizes the Glance
in a manner similar to other services to create a glance-api.conf
file with auth tokens on the controller node.