Added yaml file to build dependencies.
Added a custom yaml file for dependencies. It's meant to minimize the
amount of hard-coded information and infer as much as possible to build
out different testing scenarios.

Dependencies are listed by namespace, with options to override the
forge_namespace and the git_prefix (most projects use puppet-*,
puppetlabs projects use puppetlabs-*).

The intent is to generate a puppetfile with either github locations
or forge locations based on environment variable settings. Refs allow
for stable pinning of external dependencies.