Remove dependency on puppet-tempest
Puppet-tempest does not install correctly on Puppet Enterprise >=3.3
with the puppet module tool. This commit removes the dependency on
puppet-tempest from the Modulefile and removes the default inclusion of
tempest on the allinone node.

Users wishing to use tempest can still include the tempest role on
their node. The Puppetfile will continue to depend on puppet-tempest.