Rewrite the rabbitmq profile
nova::rabbitmq should be going away soon, so it is best to stop using
it. Moreover, the nova::rabbitmq class depends on an old version of the
rabbitmq module that contains what in some contexts is considered a
security vulnerability. While we can't change the rabbitmq version in
the metadata.json file because it will cause a conflict with the other
openstack modules, we can update our code and Puppetfile to use the
newer version and its new dependencies while maintaining compatibility
with the old version.

This commit also adds a provisioning step to the control node to ensure
that the node's hostname points to 127.0.0.1. This allows rabbitmq to
restart without being blocked by the firewall.