Remove Exec defaults from neutron router profile
Currently there is an Exec resource default to specify a global exec
path and require a non-existent profile class. There are no exec
resources in this profile for this resource default to apply to, and it
leaks into execs to which it doesn't apply. Removing it causes no
errors during a puppet run, so this commit removes it.