Replace *_address params for keystone::endpoint
The public_address, internal_address, and admin_address parameters were
removed in keystone at 29b68753316c7ceeb0a6ece98b686c216968a3dd. This
commit replaces our usage of those parameters with the new *_url
parameters, using the old default scheme, port, and path to maintain
current functionality.

Eventually this should be replaced by configurable URL parameters.
However, the address parameters are scattered throughout the
module, care needs to be taken to address this with consistency. This
patch fixes it just enough so that it compiles when using the master
branch of the modules.