Don't require that OVS agent run on neutron server
Without this patch, the neutron::agents::ml2::ovs class was bundled
into the openstack::common::ovs class, which was included in the
neutron server profile. Since the OVS agent is not required to run on
the neutron server, include it only where it's actually needed.

To accomplish this, this patch reorganizes the openstack::common::ovs
class into two classes, openstack::common::ml2 for the ml2 plugin
class, and openstack::common::ml2::ovs for the ml2 ovs agent, and
includes openstack::common::ml2 where openstack::common::ovs used to
be, but only includes the openstack::common::ml2::ovs class where
needed.

Closes #107