Reorganize yum_refresh
Without this patch, the yum_refresh class was only added for
CentOS/RedHat systems and not Fedora systems because it was being
included in the same class as epel. Since yum_refresh has nothing to do
with epel and since yum_refresh is applicable to Fedora, this patch
moves the yum_refresh inclusion to the repo class and does ordering
within the yum_refresh class instead of for just the rdo repo.