Use Puppet Labs CentOS 7 vagrant box in examples
The Puppet Labs has multiple providers, so it is more accessible for
trying out the examples.

We're pinning the version to 1.0.0 because 1.0.1 seems to have issues
with its networking interfaces.

Additionally, this vagrant box comes with firewalld enabled and many
firewall rules. Firewalld responds to iptables-save which confuses the
firewall iptables provider and causes errors on the first round of
puppet runs, even if disabling firewalld is puppet's first change.
Because of this, we add a command to the deployment script to stop
firewalld before running puppet so that puppet succeeds on the first
run.