Remove conditional to stop swift proxy service
The defined function is not able to detect whether any resources have
been collected in order to conditionally turn on the swift-proxy
service once the ring resources exist, which was the intention of this
section. The defined function instead is only able to check if the type
is loadable, and always returned true since the ring resources are
valid types. This commit removes that section since it was just a noop.