Install neutron-server on non-controller nodes
We declare the neutron::server class on non-controller nodes in order
to manage some configs needed by agents and routers. We only turn the
neutron-server service on if the node is a controller. Previously, we
avoided installing the package if the node was not a controller.
However, on Ubuntu, the neutron::plugin::ml2 class needs to manage
/etc/default/neutron-server, which comes from the neutron-server
package. Also on Ubuntu, the neutron::server class tries to manage the
neutron-server. Even if it is ensuring it is stopped, it fails if there
is no startup script for the service. This should be better architected
to avoid these issues, but the quick fix for now is to add the package
back onto the nodes by using the default package_ensure parameter for
the neutron::server class. We still ensure that the service is stopped.