Remove openstack::config class
We can reduce a lot of code duplication by setting all parameters in
the openstack class instead of passing them all through to the
openstack::config class. We also do not need to repeat the parameter
list twice in the openstack class if we set correct defaults in the
parameter list. In this patch, we set the default value of each
parameter to be an explicit hiera lookup using the hiera function, with
the default of the hiera call being the original default values. Users
need Puppet 3 in order to use the hiera call, but they are not forced
to use hiera because they can still override the values by passing them
in directly.