Remove openstack::setup::cirros class
Having a class that sets up a single hard-coded image is not very
useful, except for demos. This change moves the image data into the
example hiera files and uses create_resources on a hash of data to
create as many images as the user wants. We move this into
openstack::profile::glance::api since now a one-line function does not
need its own class.