Fix puppetlabs/mysql dependency
The stackforge/puppet-openstacklib module has updated its dependency on
puppetlabs-mysql to 3.x [1]. We do not need to have puppetlabs-mysql as
a dependency of this module as all the stackforge modules will pull
puppet-openstacklib in as a dependency, which pulls in
puppetlabs-mysql as a dependency.

This commit also pins puppetlabs-mysql to a stable 3.x branch in the
Puppetfile.

[1] https://review.openstack.org/#/c/149508/

Closes #187